Existing Client Profile Check and Redirect Complete
Using the profileUpdate.html as the display info if exists.
Created the views, new form, and html to handle client profile redirecting depending on existing user data.
If Edit button is hit, the record for that user's client profile is deleted and the user is brought to the creation of the client profile (client_profile.html)