Added quote_redirect check to views
So since we are saving quotes and displaying them in quote_redirect for user accept/ deny. We have to delete said quote if the user tries to move to any other page. I added a check to every view, where it would check if last page visited was "quote_redirect". If YES, delete last entry (NEW ONE); NO, continue on as normal.